core: Decouple bonded from non-bonded interactions
For better separation of concerns, move the bonded interactions
utility functions out of the non-bonded interactions module.
