+++ src/core/event.cpp
@@ -42,6 +42,7 @@
 #include "grid_based_algorithms/lb_interface.hpp"
 #include "immersed_boundaries.hpp"
 #include "integrate.hpp"
+#include "interactions.hpp"
 #include "nonbonded_interactions/nonbonded_interaction_data.hpp"
 #include "npt.hpp"
 #include "partCfg_global.hpp"
@@ -108,7 +109,7 @@ void on_integration_start(double time_step) {
 #ifdef NPT
   integrator_npt_sanity_checks();
 #endif
-  if (!interactions_sanity_checks()) {
+  if (long_range_interactions_sanity_checks()) {
     runtimeErrorMsg() << "Long-range interactions returned an error.";
   }
   lb_lbfluid_sanity_checks(time_step);
