+++ src/core/pressure.cpp
@@ -32,6 +32,7 @@
 #include "config.hpp"
 #include "event.hpp"
 #include "grid.hpp"
+#include "interactions.hpp"
 #include "nonbonded_interactions/nonbonded_interaction_data.hpp"
 #include "pressure_inline.hpp"
 #include "virtual_sites.hpp"
@@ -56,7 +57,7 @@ static std::shared_ptr<Observable_stat> calculate_pressure_local() {
 
   auto obs_pressure_ptr = std::make_shared<Observable_stat>(9);
 
-  if (!interactions_sanity_checks())
+  if (long_range_interactions_sanity_checks())
     return obs_pressure_ptr;
 
   auto &obs_pressure = *obs_pressure_ptr;
